# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Outbound dialer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TwilioStatus {
  connected: boolean;
  phoneNumber?: string | null;
  message?: string | null;
}

export interface Lead {
  id: number;
  name: string;
  phone?: string | null;
  address?: string | null;
  website?: string | null;
  rating?: number | null;
  reviewCount?: number | null;
  signalTags?: string | null;
  latitude?: number | null;
  longitude?: number | null;
  placeId?: string | null;
  leadScore: number;
  tier: string;
  status: string;
  notes?: string | null;
  lastDisposition?: string | null;
  lastCalledAt?: string | null;
  callCount?: number;
  createdAt: string;
}

export interface Call {
  id: number;
  leadId: number;
  scriptId?: number | null;
  campaignId?: number | null;
  twilioCallSid?: string | null;
  status: string;
  disposition?: string | null;
  notes?: string | null;
  durationSec?: number | null;
  pathTaken?: string | null;
  startedAt?: string | null;
  endedAt?: string | null;
  createdAt: string;
}

export type LeadDetail = Lead & {
  calls?: Call[];
};

export interface LeadList {
  items: Lead[];
  total: number;
}

export interface CreateLeadBody {
  name: string;
  phone?: string | null;
  address?: string | null;
  website?: string | null;
  leadScore?: number;
  tier?: string;
  status?: string;
  notes?: string | null;
}

export interface UpdateLeadBody {
  name?: string;
  phone?: string | null;
  address?: string | null;
  website?: string | null;
  leadScore?: number;
  tier?: string;
  status?: string;
  notes?: string | null;
  lastDisposition?: string | null;
}

export type LeadStatsByStatusItem = {
  key: string;
  count: number;
};

export type LeadStatsByTierItem = {
  key: string;
  count: number;
};

export type LeadStatsByDispositionItem = {
  key: string;
  count: number;
};

export interface LeadStats {
  total: number;
  withPhone: number;
  called: number;
  hotLeads?: number;
  byStatus: LeadStatsByStatusItem[];
  byTier: LeadStatsByTierItem[];
  byDisposition: LeadStatsByDispositionItem[];
}

export interface ScriptSummary {
  id: number;
  name: string;
  description?: string | null;
  isActive: boolean;
  nodeCount: number;
  rootNodeId?: number | null;
  createdAt: string;
}

export interface ScriptOption {
  key: string;
  label: string;
  nextNodeId?: number | null;
  disposition?: string | null;
}

export interface ScriptNode {
  id: number;
  scriptId: number;
  title?: string | null;
  message: string;
  nodeType?: string;
  options: ScriptOption[];
}

export interface Script {
  id: number;
  name: string;
  description?: string | null;
  isActive: boolean;
  rootNodeId?: number | null;
  createdAt: string;
  nodes: ScriptNode[];
}

export interface CreateScriptBody {
  name: string;
  description?: string | null;
  isActive?: boolean;
}

export interface UpdateScriptBody {
  name?: string;
  description?: string | null;
  isActive?: boolean;
  rootNodeId?: number | null;
}

export interface CreateScriptNodeBody {
  title?: string | null;
  message: string;
  nodeType?: string;
  options?: ScriptOption[];
}

export interface UpdateScriptNodeBody {
  title?: string | null;
  message?: string;
  nodeType?: string;
  options?: ScriptOption[];
}

export interface Campaign {
  id: number;
  name: string;
  scriptId: number;
  status: string;
  leadFilter?: string | null;
  callsMade?: number;
  createdAt: string;
}

export type CampaignDetail = Campaign & {
  script?: Script;
  queueSize?: number;
};

export interface CreateCampaignBody {
  name: string;
  scriptId: number;
  leadFilter?: string | null;
}

export interface NextLeadResponse {
  lead?: Lead;
  remaining?: number;
}

export type CallWithLead = Call & {
  leadName?: string;
  leadPhone?: string | null;
};

export interface StartCallBody {
  leadId: number;
  scriptId?: number | null;
  campaignId?: number | null;
}

export interface UpdateCallBody {
  status?: string;
  disposition?: string | null;
  notes?: string | null;
  pathTaken?: string | null;
}

export type DashboardSummaryDispositionsItem = {
  key: string;
  count: number;
};

export interface DashboardSummary {
  totalLeads: number;
  leadsWithPhone: number;
  totalCalls: number;
  callsToday: number;
  hotLeads: number;
  connectedToday?: number;
  avgDurationSec?: number;
  dispositions: DashboardSummaryDispositionsItem[];
}

export type GetRecentCallsParams = {
  limit?: number;
};

export type ListLeadsParams = {
  search?: string;
  status?: string;
  tier?: string;
  hasPhone?: boolean;
  limit?: number;
  offset?: number;
};

export type ListCallsParams = {
  leadId?: number;
  limit?: number;
};