# lib/api-zod/src/generated/types/call.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Outbound dialer API
 * OpenAPI spec version: 0.1.0
 */

export interface Call {
  id: number;
  leadId: number;
  scriptId?: number | null;
  campaignId?: number | null;
  twilioCallSid?: string | null;
  status: string;
  disposition?: string | null;
  notes?: string | null;
  durationSec?: number | null;
  pathTaken?: string | null;
  startedAt?: Date | null;
  endedAt?: Date | null;
  createdAt: Date;
}