# lib/api-zod/src/generated/types/callWithLead.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Outbound dialer API
 * OpenAPI spec version: 0.1.0
 */
import type { Call } from "./call";

export type CallWithLead = Call & {
  leadName?: string;
  leadPhone?: string | null;
};