# lib/api-zod/src/generated/types/nextLeadResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Outbound dialer API
 * OpenAPI spec version: 0.1.0
 */
import type { Lead } from "./lead";

export interface NextLeadResponse {
  lead?: Lead;
  remaining?: number;
}